Sprites imported from other branch; TheBox refinement; Player now explicitly turns around
The Box and the Player now explicitly turn around, allowing us to flip the sprites and move any hitboxes later on.

The Box is now a platform when it isn't being carried.

The Box now gets thrown to the left or the right depending on if the player is facing the right way.

The player can now throw The Box. The Box has been added to PlayerTester to demonstrate this.